+++ docs/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 Things under "main" are not released yet.
 
+## main
+
+- Allow `std.isFunction` and similar to learn that variables have function type.
+- Add more examples to std lib docs.
+
 ## v0.7.2
 
 - Improve go-to-def, etc through asserts.